Refatoring, separate svg calls to allow testing
